@@ -7,17 +7,15 @@
     "license": "MIT",
     "require": {
         "php-amqplib/php-amqplib": "^3.1.0",
-        "symfony/options-resolver": "^2.4 || ^3.0 || ^4.0 || ^5.0",
+        "symfony/options-resolver": "^6.2",
         "psr/log": "*"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5.10",
-        "gonzalo123/yaml2pimple": "*",
         "phpmd/phpmd": "^2.6",
         "friendsofphp/php-cs-fixer": "^3.2.1"
     },
     "suggest": {
-        "gonzalo123/yaml2pimple": "Create connection object for you",
         "symfony/symfony": "Great framework with container able to create connection object for you"
     },
     "autoload": {
