@@ -11,17 +11,17 @@ use PhpAmqpLib\Message\AMQPMessage;
 class ConsumerCallback
 {
     private $callback;
-    private $MessageSerializer;
+    private $messageSerializer;
 
     public function __construct(callable $callback, MessageSerializer $messageSerializer)
     {
         $this->callback = $callback;
-        $this->MessageSerializer = $messageSerializer;
+        $this->messageSerializer = $messageSerializer;
     }
 
     public function __invoke(AMQPMessage $msg)
     {
-        $result = \call_user_func($this->callback, new Message($this->MessageSerializer->deserialize($msg->body)));
+        $result = \call_user_func($this->callback, new Message($this->messageSerializer->deserialize($msg->body)));
         if ($result === true) {
             $msg->delivery_info['channel']->basic_ack($msg->delivery_info['delivery_tag']);
         } else {
